dfSDfasdf

+++ controllers/videoController.js
@@ -31,8 +31,11 @@ export const PostUpload = async (req,res) => {
     const newVideo = await Video.create({
         fileUrl:path,
         title,
-        description
+        description,
+        creator: req.user.id
     })
+    req.user.videos.push(newVideo._id);
+    req.user.save();
     //To Do: Upload and save Video
     res.redirect(routes.videoDetail(newVideo.id));
 };
@@ -42,7 +45,7 @@ export const videoDetail = async(req,res) => {
         params:{id}
     } = req;
     try{
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate("creator");
         res.render("videoDetail", {pageTitle : video.title,video})
     }catch(error){
         res.redirect(routes.home);
@@ -54,7 +57,11 @@ export const getEditVideo = async(req,res) => {
     } = req;
     try{
         const video = await Video.findById(id);
-        res.render("editVideo",{pageTitle:`Edit ${video.title}`,video})
+        if(String(video.creator) !== req.user.id){
+            throw Error();
+        }else{
+            res.render("editVideo",{pageTitle:`Edit ${video.title}`,video})
+        }
     }
     catch(error){
         res.redirect(routes.home);
@@ -77,7 +84,12 @@ export const deleteVideo = async(req,res) => {
         params:{id}
     } = req;
     try{
-        await Video.findOneAndRemove({_id:id});
+        const video = await Video.findById(id);
+        if(String(video.creator) !== req.user.id){
+            throw Error();
+        }else{
+            await Video.findOneAndRemove({_id:id});
+        }
     }catch{}
     res.redirect(routes.home);
 };
